+++ app-modules/knowledge-base/src/Filament/Resources/KnowledgeBaseItems/Pages/ListKnowledgeBaseItems.php
@@ -161,7 +161,6 @@ class ListKnowledgeBaseItems extends ListRecords
                     }),
             ], layout: FiltersLayout::BeforeContent)
             ->recordActions([
-                EditAction::make(),
                 ReplicateAction::make()
                     ->schema([
                         Section::make()
